test(bpp): de-duplicate withEnv helper to satisfy SonarCloud new-code gate
SonarCloud flagged 5.7% duplication on new code (gate ≤ 3%), all from the
identical ~29-line `withEnv` env-override helper copy-pasted across three new
test files (usage-ingest.test.ts, usage-push-job.test.ts, and the usage-push
cron route test).

Extract it once into src/test-utils/env.ts and import it in all three. No
behavior, signature, flag-default, or security-property change — pure test
refactor. Lint/tsc/build clean; test counts unchanged (204 pass, 18 DB-skipped).
