+++ src/app/api/v1/internal/usage-push/route.test.ts
@@ -3,37 +3,9 @@ import assert from "node:assert/strict";
 
 import { NextRequest } from "next/server";
 
-const CRON_SECRET = "cron-secret-value";
+import { withEnv } from "@/test-utils/env";
 
-function withEnv(
-  overrides: Record<string, string | undefined>,
-  fn: () => Promise<void> | void,
-): Promise<void> | void {
-  const keys = Object.keys(overrides);
-  const previous = new Map(keys.map((k) => [k, process.env[k]]));
-  for (const [k, v] of Object.entries(overrides)) {
-    if (v === undefined) delete process.env[k];
-    else process.env[k] = v;
-  }
-  const restore = () => {
-    for (const k of keys) {
-      const prev = previous.get(k);
-      if (prev === undefined) delete process.env[k];
-      else process.env[k] = prev;
-    }
-  };
-  try {
-    const result = fn();
-    if (result instanceof Promise) {
-      return result.finally(restore);
-    }
-    restore();
-    return result;
-  } catch (error) {
-    restore();
-    throw error;
-  }
-}
+const CRON_SECRET = "cron-secret-value";
 
 function cronRequest(token?: string): NextRequest {
   const headers: Record<string, string> = {};
